moved optimal grouping to algorithms
